+++ NEWS.md
@@ -1,3 +1,13 @@
+# cinar 0.2.0
+* Added:
+  - `heatmap_differential` function to be able to plot the heatmaps
+of a given comparison.
+  - `show_comparisons` is added to see the available comparisons 
+before you plot differential heatmaps.
+
+* Changed:
+  - `heatmap_plot` is now renamed to `heatmap_var_peaks`, which makes more sense.
+
 # cinaR 0.1.1
 This version is even cooler, lol: